updated img to bin

+++ img_to_bin.py
@@ -9,10 +9,10 @@ rootdir = sys.argv[1]
 k = 0
 
 for imgdir in os.listdir("./" + rootdir):
-    k += 1 
+    k += 1
 
-rgb_arr = np.zeros([k, 256, 256, 3], dtype="int8")
-gray_arr = np.zeros([k, 256, 256], dtype="int8")
+rgb_arr = np.zeros([k, 256, 256, 3], dtype="int32")
+gray_arr = np.zeros([k, 256, 256], dtype="int32")
 
 k = 0
 for imgdir in os.listdir("./" + rootdir):
@@ -22,9 +22,5 @@ for imgdir in os.listdir("./" + rootdir):
     gray_arr[k] = np.asarray(img.convert('L'))
     k += 1
 
-np.save(rootdir + "_rgb" , rgb_arr)
-np.save(rootdir + "_gray" , gray_arr)
-
-
-
-
+np.save(rootdir + "_rgb", rgb_arr)
+np.save(rootdir + "_gray", gray_arr)
